Silverlight on Windows Phone does not support atomic operations on 64-bit integers: http://msdn.microsoft.com/en-us/library/zs86dyzy(v=vs.95).aspx

+++ ReactiveUI/ObservableAsyncMRUCache.cs
@@ -197,7 +197,7 @@ namespace ReactiveUI
         static readonly Logger log = LogManager.GetCurrentClassLogger();
         readonly ISubject<T> _inner;
         Queue<T> _nextItems = new Queue<T>();
-        long _count;
+        int _count;
         readonly int _maxCount;
 
         public SemaphoreSubject(int maxCount, IScheduler sched = null)
